requirements.txt
Flask==0.10.1
Flask-Gravatar==0.4.1
Flask-Navigation==0.2.0
Flask-Stormpath==0.2.9
Jinja2==2.7.3
MarkupSafe==0.23
PyJWT==0.3.0
Werkzeug==0.9.6
httplib2==0.9
itsdangerous==0.24
python-dateutil==2.2
requests==2.4.3
requests-oauthlib==0.4.2
stormpath==1.2.4
